requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
fastapi[standard]
psycopg2
psycopg2-binary
pandas